Restructure thrFiringSchmittAlg Types.h to pure C
